requirements-windows.txt
# Windows Receiver Requirements
# Python 3.12+ required

Pillow>=12.0.0
pyperclip>=1.11.0
cryptography>=42.0.0
paddleocr>=2.8.0
paddlepaddle>=2.6.0
protobuf==3.20.2
